refactor: make encryption/dal internal
This continues the service split refactoring.
